% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Retrieve the available LINCS datasets and the corresponding assay type.}
\usage{
available_data()
}
\value{
data frame with 2 columns, LDS-ID and assay name.
}
\description{
Retrieve the available LINCS datasets and the corresponding assay type.
}
\examples{
available_data()
}
