% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gcp_expression.R
\name{get_gcp_expression}
\alias{get_gcp_expression}
\title{Retrieve Broad Global Chromatin Profiling (GCP) data, plate normalized log2 fold change.}
\usage{
get_gcp_expression(filter_foldchange = "none", histone_modification = "all",
  cells = "all", perturbagens = "all", perturbagen_dose = "all",
  replicates = "all", limits = 1000)
}
\arguments{
\item{filter_foldchange}{string expressing a threshold for specific log2 fold change retrieval (default "all", admissible ">value" or "<value")}

\item{histone_modification}{vector of strings with histone modifications (Default "all")}

\item{cells}{vector of strings with cell names (Default "all")}

\item{perturbagens}{vector of strings with perturbagens IDs (Default "all")}

\item{perturbagen_dose}{numeric vector with perturbagen doses (Default "all")}

\item{replicates}{string, return each replicate separately (Default "all") or aggregated as averge value (using replicates = "mean")}

\item{limits}{numeric value, return only the specified number of rows (Default 1000)}

\item{perturbagen_time}{numeric vector with perturbagen times (Default "all")}
}
\value{
data frame containing entity selected and the associated plate normalized log2 fold change
}
\description{
Retrieve Broad Global Chromatin Profiling (GCP) data, plate normalized log2 fold change.
}
\examples{
get_gcp_expression(cells = c("MCF7","PC3"))
get_gcp_expression(cells = c("MCF7","PC3"),filter_zscores=">2")
get_gcp_expression(perturbagens = "BRD-K31283835")
get_gcp_expression(histone_modification = "acetylation")
}
