% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imaging.R
\name{get_imaging}
\alias{get_imaging}
\title{Retrieve Imaging assays data}
\usage{
get_imaging(lds_id = NULL, cells = "all", perturbagens = "all")
}
\arguments{
\item{lds_id}{string equal to a dataset id, no default}

\item{cells}{vector of strings with protein names (Default "all")}

\item{perturbagens}{vector of strings with perturbagens IDs (Default "all")}
}
\value{
data frame containing imaging data from the specified LDS-ID
}
\description{
Retrieve Imaging assays data
}
\examples{
get_imaging(lds_id="LDS-1263")
get_imaging(lds_id="LDS-1263",perturbagens="10049-101-1")
get_imaging(lds_id="LDS-1221",cells="C32")
}
