% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinativ.R
\name{get_kinativ}
\alias{get_kinativ}
\title{Retrieve KiNativ assay data, percent inhibition or ic50.}
\usage{
get_kinativ(level = "perc_inhib", cells = "all", proteins = "all",
  perturbagens = "all", perturbagen_dose = "all")
}
\arguments{
\item{level}{string equal to "perc_inhib" or "ic50" for percent inhibition or ic50 data respectively (default "perc_inhib")}

\item{proteins}{vector of strings with protein names (Default "all")}

\item{perturbagens}{vector of strings with perturbagens IDs (Default "all")}

\item{perturbagen_dose}{numeric vector with perturbagen doses (Default "all")}
}
\value{
data frame containing entity selected and the associated percent inhibition or ic50 data
}
\description{
Retrieve KiNativ assay data, percent inhibition or ic50.
}
\examples{
get_kinativ(level="perc_inhib")
get_kinativ(level="ic50",proteins="BRAF")
get_kinativ(perturbagens = "10231-101-1")
}
