% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kinomescan.R
\name{get_kinomescan}
\alias{get_kinomescan}
\title{Retrieve KINOMEscan assay data, percent control or kd.}
\usage{
get_kinomescan(level = "perc_ctrl", filter_perc_or_kd = "none",
  proteins = "all", perturbagens = "all", perturbagen_dose = "all")
}
\arguments{
\item{level}{string equal to "perc_ctrl" or "kd" for percent control or kd data respectively (default "perc_ctrl")}

\item{filter_perc_or_kd}{string expressing a threshold for specific percent inhibition or kd retrieval (default "none")}

\item{proteins}{vector of strings with protein names (Default "all")}

\item{perturbagens}{vector of strings with perturbagens IDs (Default "all")}

\item{perturbagen_dose}{numeric vector with perturbagen doses (Default "all")}
}
\value{
data frame containing entity selected and the associated percent control or kd data
}
\description{
Retrieve KINOMEscan assay data, percent control or kd.
}
\examples{
get_kinomescan(level="perc_ctrl")
get_kinomescan(level="kd",proteins="BRAF")
get_kinomescan(perturbagens = "10083-101-1",filter_perc_or_kd="<10")
}
