% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_L1000ph2_expression.R
\name{get_L1000ph2_expression}
\alias{get_L1000ph2_expression}
\title{Retrieve Broad L1000 data from LINCS Phase 2, weighted average of replicates z-scores}
\usage{
get_L1000ph2_expression(sig_ids = "all", filter_zscores = "none",
  genes = "all", cells = "all", perturbagens = "all",
  perturbagen_type = "all", perturbagen_dose = "all",
  perturbagen_time = "all", limits = 1000)
}
\arguments{
\item{filter_zscores}{string expressing a threshold for specific z-scores retrieval (default "all", admissible ">value" or "<value")}

\item{genes}{vector of strings with gene names (Default "all")}

\item{cells}{vector of strings with cell names (Default "all")}

\item{perturbagens}{vector of strings with perturbagens IDs (Default "all")}

\item{perturbagen_type}{vector of strings with perturbagen types (Default "all")}

\item{perturbagen_dose}{numeric vector with perturbagen doses (Default "all")}

\item{perturbagen_time}{numeric vector with perturbagen times (Default "all")}

\item{limits}{numeric value, return only the specified number of rows (Default 1000)}
}
\value{
data frame containing entity selected and the associated wegthted average z-scores
}
\description{
Retrieve Broad L1000 data from LINCS Phase 2, weighted average of replicates z-scores
}
\examples{
get_L1000ph2_expression(genes = c("EGF","RHEB"))
get_L1000ph2_expression(cells = c("MCF7","PC3"),filter_zscores=">2")
get_L1000ph2_expression(cells = c("MCF7","PC3"),filter_zscores=">2")
get_L1000ph2_expression(perturbagens = "BRD-K60230970-001-10-0")
}
