% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mema.R
\name{get_mema}
\alias{get_mema}
\title{Retrieve MEMA assay data, for different staining sets.}
\usage{
get_mema(ligands = "all", ecmps = "all", cells = "all",
  staining_set = 1, limits = 100)
}
\arguments{
\item{ligands}{vector of strings with ligand names (Default "all")}

\item{ecmps}{vector of strings with extracellular matrix proteins (Default "all")}

\item{cells}{vector of strings with cell names (Default "all")}

\item{staining_set}{integer for the different staining sets (1 to 4 for staining set 1, 2, 3 or 2noH3 respectively)}
}
\value{
data frame containing combination of perturbations and a set of measured features
}
\description{
Retrieve MEMA assay data, for different staining sets.
}
\examples{
get_mema(staining_set=2)
get_mema(cells="PC3")
get_mema(staining_set=3, ligands="BMP3_P12645")
}
