% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rppa.R
\name{get_rppa}
\alias{get_rppa}
\title{Retrieve RPPA assays data}
\usage{
get_rppa(lds_id = NULL, cells = "all", perturbagens = "all")
}
\arguments{
\item{lds_id}{string equal to a dataset id, no default}

\item{cells}{vector of strings with protein names (Default "all")}

\item{perturbagens}{vector of strings with perturbagens IDs (Default "all")}
}
\value{
data frame containing RPPA data from the specified LDS-ID
}
\description{
Retrieve RPPA assays data
}
\examples{
get_rppa(lds_id="LDS-1209")
get_rppa(lds_id="LDS-1209",cells="C32")
}
